Get rid of `ApplyT` in `configureFakeintake`
We need to get rid of `ApplyT` because it was breaking the dependency
between the fake intake ECS task and the Helm release.
Concretely, when enabling the fake intake, the first `pulumi up` was
only spawning the fake intake and a second `pulumi up` was needed to
have the agent configuration updated with the fake intake endpoint.

The `ApplyT` was needed to be able to use the fake intake
IP (`StringInput` type) as a `pulumi.Map` key (`string` type).

Using environment variables instead permits to need the fake intake IP
only in `pulumi.Map` values (`Input` type).