Target testing keys directory when downloading artifacts from {apt,yum}testing.datad0g.com (#1652)
When targeting a test artifact from a specific pipeline in datadog-agent, use the testing keys to verify the metadata signatures of apttesting.datad0g.com and yumtesting.datad0g.com.

Needed after the merge of DataDog/datadog-agent#36943, which makes the datadog-agent pipeline sign testing repositories with a dedicated set of testing keys.