CONTINT-4687 - add fakeintake options
Add necessary code to pass option to start fakeintake server:

- `ddagent:fakeintakeStoreType` to select the store tyep (sql or memory)
- `ddagent:fakeintakeRetentionPeriod` to select the data retention
  period

The `fakeintakeRetentionPeriod` is useful for the coming end2end tests
to ensure we store the data we want to test long enough for the test to
run and prevent having the scenario where the test runs and the data we
request has been cleaned-up.

Added the argument parsing for all scenarios that use the fakeintake and
allows it to be configured.

Signed-off-by: Alexandre Lavigne <alexandre.lavigne@datadoghq.com>