CONTINT-4792 - stress-ng app use tmp directory
the stress-ng app requires it's current working directory to be writable
and a given temporary path to be writable.

currently we provide no temporary path so stress-ng uses its current
working direcotyr path.
see here when option `temp-path` is *not* given:
https://github.com/ColinIanKing/stress-ng/blob/85e2ebc2bb3c1aa2b8db767c17fa4ed829a3f1c6/core-filesystem.c#L356-L358

To solve that:

1. request a temporary directory in the pod spec under the path `/tmp`
2. pass that temporary directory to stress-ng
3. change the working directory too for io-stress to work too.

We need to way to identify that we run an openshit scenario and enable
the nested virtualization.

Use a instance type that allows nested virtualization. google cloud docs
entions that all `e2` instances can't run nested virt.

to conclude, use the regular stress-ng deployment and remove the
openshift dedicated one.

Signed-off-by: Alexandre Lavigne <alexandre.lavigne@datadoghq.com>