CONTINT-4792 - on openshift attach serviceAccount to tracegen app
Refactor the function to deploy tracegen application.

On openshift, we must create a new service account that allows us to
have some priviledged access to the host for the app tracegen-uds.

In openshift scenario: create the serviceAccount, create the
clusterRoleBinding to bind that service account to the priviledged
SecurityContextConstraint from openshift.

Signed-off-by: Alexandre Lavigne <alexandre.lavigne@datadoghq.com>